+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Change Log
 vUNRELEASED
 ---------------
 * :+1: Added the possibility to import `Parts` from an `Excel` file (#931)
+* :+1: Added the possibility to clone `Widgets` from one `Form status` to another (#1219)
 
 v4.3.0 (7NOV22)
 ---------------